lib: Allow the OS to map the regions that are protected using PMP
This is achieved by removing the 'no-map' property from the
'reserved-memory' node.

Telling the OS not to map the reserved regions does not add much
protection since it only avoids access to regions that are already
protected by PMP. But by not allowing the OS to map those regions, it
creates holes in the OS system memory map and prevents the use of
hugepages which would generate, among other benefits, less TLB miss.

Signed-off-by: Alexandre Ghiti <alex@ghiti.fr>